refactor(summary-notification): rename indicator_object.py to ops_indicators.py
ユビキタス言語「運用指標 = ops indicators」に揃える。
クラス名 OpsIndicators は変更なし。テストファイルも追従リネーム。
