+++ lambda/summary-notification/src/domain/ops_indicators.py
