+++ lambda/summary-notification/tests/domain/test_ops_indicators.py
